+++ securedrop/tests/pages-layout/functional_test.py
@@ -20,9 +20,6 @@ import io
 from os.path import abspath, dirname, realpath
 import pytest
 
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.common.keys import Keys
-
 from selenium import webdriver
 
 from tests.functional import functional_test
@@ -46,33 +43,13 @@ class FunctionalTest(functional_test.FunctionalTest):
                          'screenshots', self.accept_languages))
         if not os.path.exists(self.log_dir):
             os.makedirs(self.log_dir)
-        firefox = self._prepare_webdriver()
-        profile = webdriver.FirefoxProfile()
-        profile.set_preference("intl.accept_languages", self.accept_languages)
+        self.new_profile = webdriver.FirefoxProfile()
+        self.new_profile.set_preference("intl.accept_languages", self.accept_languages)
         self.override_driver = True
-        self.driver = self._create_webdriver(firefox, profile)
-        self._javascript_toggle()
-
         yield None
 
-        self.driver.quit()
-
     def _screenshot(self, filename):
-        self.driver.set_window_size(1024, 500)  # Trim size of images for docs
         self.driver.save_screenshot(os.path.join(self.log_dir, filename))
-
-    def _javascript_toggle(self):
-        # the following is a noop for some reason, workaround it
-        # profile.set_preference("javascript.enabled", False)
-        # https://stackoverflow.com/a/36782979/837471
-        self.driver.get("about:config")
-        actions = ActionChains(self.driver)
-        actions.send_keys(Keys.RETURN)
-        actions.send_keys("javascript.enabled")
-        actions.perform()
-        actions.send_keys(Keys.TAB)
-        actions.send_keys(Keys.RETURN)
-        actions.perform()
 
     def _save_alert(self, filename):
         fd = io.open(os.path.join(self.log_dir, filename), 'wb')
